fix bug with state being refreshed to initialState when tournament show page is refreshed, build out skeleton tournament detail, begin refactoring tournament detail into multiple components
