+++ frontend/src/components/tournamentComponents/tournamentShow/index.js
@@ -1,21 +1,29 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { fetchTournament } from '../../../actions/tournamentActions';
+import { fetchTournament, fetchTournaments } from '../../../actions/tournamentActions';
 
 //components
-import TournamentDetail from './TournamentDetail';
+import TournamentDetail from './tournamentDetail';
 import ForumButton from './forumButton';
 
 class TournamentShow extends Component {
+
   componentDidMount() {
+    this.props.fetchTournaments();
     this.props.fetchTournament(this.props.id);
   }
 
   render() {
     const { selectedTournament } = this.props;
 
+    if (!selectedTournament) {
+      return <div className="center brand-logo" style={{height: "20em", padding: "50px", fontSize: "24px"}}>
+        Loading Tournament...
+      </div>
+    }
+
     return (
-      <div className="container">
+      <div className="container" style={{height: "650px"}}>
         <h5 className="brand-logo">TournamentShow</h5>
         <TournamentDetail tournament={selectedTournament}/>
         <ForumButton/>
@@ -29,4 +37,4 @@ const mapStateToProps = (state, ownProps) => ({
   id: ownProps.match.params.id
 })
 
-export default connect(mapStateToProps, { fetchTournament })(TournamentShow);
+export default connect(mapStateToProps, { fetchTournament, fetchTournaments })(TournamentShow);
